double to foat ostick 1000→10000

+++ Core/Inc/motor.h
@@ -10,8 +10,6 @@
 
 #include "main.h"
 
-
-
 //motor
 #define MR 0
 #define ML 1
@@ -19,12 +17,8 @@
 #define STEPPER_CLOCK_R_CHANNEL TIM_CHANNEL_1
 #define STEPPER_CLOCK_L_CHANNEL TIM_CHANNEL_1
 
-
-
-
-
-uint32_t SPEEDtoHz(uint32_t speed);
-uint32_t HztoSPEED(uint32_t Hz);
-void mortor_direction(uint8_t motor,uint8_t direction);
+float SPEEDtoHz(float speed);
+float HztoSPEED(float Hz);
+void mortor_direction(uint8_t motor, uint8_t direction);
 void mortor_stop(void);
 #endif /* INC_MOTOR_H_ */
